split up enricher into lib and boudicca.events enricher
