+++ .run/LocalEnricher.run.xml
@@ -1,7 +1,7 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="LocalEnricher" type="JetRunConfigurationType">
-    <option name="MAIN_CLASS_NAME" value="base.boudicca.enricher.EnricherApplicationKt" />
-    <module name="boudicca.boudicca.base.enricher.main" />
+    <option name="MAIN_CLASS_NAME" value="events.boudicca.enricher.EnricherApplicationKt" />
+    <module name="boudicca.boudicca.events.enricher.main" />
     <shortenClasspath name="NONE" />
     <extension name="net.ashald.envfile">
       <option name="IS_ENABLED" value="false" />
