--- boudicca.base/enricher-utils/build.gradle.kts
@@ -1,9 +1,0 @@
-plugins {
-    id("boudicca-kotlin")
-}
-
-dependencies {
-    implementation(libs.json)
-    implementation(project(":boudicca.base:enricher-client"))
-    implementation(project(":boudicca.base:publisher-client"))
-}
