--- boudicca.base/enricher/src/main/kotlin/base/boudicca/enricher/service/location/LocationEnricherNoopUpdater.kt
@@ -1,5 +1,0 @@
-package base.boudicca.enricher.service.location
-
-class LocationEnricherNoopUpdater : LocationEnricherUpdater {
-    override fun updateData(): List<LocationData> = emptyList()
-}
