--- boudicca.base/enricher/src/main/kotlin/base/boudicca/enricher/service/musicbrainz/Artist.kt
@@ -1,7 +1,0 @@
-package base.boudicca.enricher.service.musicbrainz
-
-data class Artist(
-    val name: String,
-    val genre: String?,
-    val lowercaseName: String = name.lowercase(),
-)
