+++ boudicca.base/enricher/src/test/kotlin/base/boudicca/enricher/TestApplication.kt
@@ -1,0 +1,6 @@
+package base.boudicca.enricher
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+
+@SpringBootApplication
+class TestApplication
