+++ boudicca.events/enricher-utils/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+    id("boudicca-kotlin")
+}
+
+group = "events.boudicca"
+
+dependencies {
+    implementation(libs.json)
+    implementation(project(":boudicca.base:enricher-client"))
+    implementation(project(":boudicca.base:publisher-client"))
+}
