+++ boudicca.events/enricher/src/main/kotlin/events/boudicca/enricher/service/location/LocationEnricherNoopUpdater.kt
@@ -1,0 +1,5 @@
+package events.boudicca.enricher.service.location
+
+class LocationEnricherNoopUpdater : LocationEnricherUpdater {
+    override fun updateData(): List<LocationData> = emptyList()
+}
