+++ boudicca.events/enricher/src/main/kotlin/events/boudicca/enricher/service/musicbrainz/Artist.kt
@@ -1,0 +1,7 @@
+package events.boudicca.enricher.service.musicbrainz
+
+data class Artist(
+    val name: String,
+    val genre: String?,
+    val lowercaseName: String = name.lowercase(),
+)
