+++ boudicca.events/enricher/src/test/kotlin/events/boudicca/enricher/EnricherApplicationSmokeTest.kt
@@ -1,0 +1,14 @@
+package events.boudicca.enricher
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class EnricherApplicationSmokeTest {
+
+    @Test
+    fun contextLoads() {
+        // This test will automatically pass if the application context loads successfully.
+        // If there's any bean misconfiguration or other Spring-related errors, it will fail.
+    }
+}
